8.5-12.2

+++ PhotosService/Controllers/PhotosApiController.cs
@@ -1,12 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using PhotosService.Data;
 using PhotosService.Models;
+using PhotosService.Services;
 
 namespace PhotosService.Controllers
 {
@@ -25,30 +28,59 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetPhotos(string ownerId)
-        {
+        public async Task<IActionResult> GetPhotos(string ownerId, [FromHeader] JwtSecurityToken accessToken)
+        { 
+            if (accessToken.Subject != ownerId)
+                return Forbid();
             var photoEntities = (await photosRepository.GetPhotosAsync(ownerId)).ToList();
             var photos = mapper.Map<List<PhotoDto>>(photoEntities);
             foreach(var photo in photos)
                 photo.Url = GeneratePhotoUrl(photo);
             return Ok(photos.ToList());
         }
-
-        [HttpGet("{id}/meta")]
-        public async Task<IActionResult> GetPhotoMeta(Guid id)
+        
+        [AllowAnonymous]
+        [HttpGet("{id}/signed-content")]
+        public async Task<IActionResult> GetPhotoSignedContent(Guid id)
         {
+            var currentUrl = HttpContext.Request.GetEncodedUrl();
+            var check = SignedUrlHelpers.CheckSignedUrl(currentUrl);
+            if (!check)
+                return Forbid();
+    
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
 
+            var photoContent = await photosRepository.GetPhotoContentAsync(id);
+            if (photoContent == null)
+                return NotFound();
+
+            return File(photoContent.Content, photoContent.ContentType, photoContent.FileName);
+        }
+
+
+        [HttpGet("{id}/meta")]
+        public async Task<IActionResult> GetPhotoMeta(Guid id, [FromHeader] JwtSecurityToken accessToken)
+        {
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             var photo = mapper.Map<PhotoDto>(photoEntity);
             photo.Url = GeneratePhotoUrl(photo);
             return Ok(photo);
         }
 
         [HttpGet("{id}/content")]
-        public async Task<IActionResult> GetPhotoContent(Guid id)
+        public async Task<IActionResult> GetPhotoContent(Guid id, JwtSecurityToken accessToken)
         {
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             var photoContent = await photosRepository.GetPhotoContentAsync(id);
             if (photoContent == null)
                 return NotFound();
@@ -57,8 +89,10 @@ namespace PhotosService.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo)
+        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != photo.OwnerId)
+                return Forbid();
             var content = Convert.FromBase64String(photo.Base64Content);
             var result = await photosRepository.AddPhotoAsync(photo.Title, photo.OwnerId, content);
             if (!result)
@@ -67,12 +101,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo)
+        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             photoEntity.Title = photo.Title;
             var result = await photosRepository.UpdatePhotoAsync(photoEntity);
             if (!result)
@@ -81,12 +116,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeletePhoto(Guid id)
+        public async Task<IActionResult> DeletePhoto(Guid id, [FromHeader] JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
-
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
             var result = await photosRepository.DeletePhotoAsync(photoEntity);
             if (!result)
                 return Conflict();
@@ -95,11 +131,15 @@ namespace PhotosService.Controllers
 
         private string GeneratePhotoUrl(PhotoDto photo)
         {
-            var relativeUrl = Url.Action(nameof(GetPhotoContent), new {
+            var relativeUrl = Url.Action(nameof(GetPhotoSignedContent), new {
                 id = photo.Id
             });
             var url = "https://localhost:6001" + relativeUrl;
-            return url;
+
+            var nowUtc = DateTime.UtcNow;
+            var signedUrl = SignedUrlHelpers.CreateSignedUrl(url, nowUtc, nowUtc.AddMinutes(5));
+            return signedUrl;
         }
+
     }
 }
